Fixes to GLYX, MGLY uptake
Add IDT checks so that STT array is not being accessed for indices that
don't exist. Checks are separate for IDTGLYX & IDTMGLY vs IDTSOAG & IDTSOAM
so that gas-phase uptake can be computed and used in non-SOA simulations
IFF GLYX and MGLY are included in the simulation as tracers (and therefore
with a value in STT). Otherwise there is no point in including these
routines as neither aerosol nor gas phase species will be updated!

Also cleaned this up a little and commented out some print statements.

Signed-off-by: Jenny Fisher <jennyf@uow.edu.au>